Add proxy support for downloading
This add support for downloading with proxies as often found in
corporate settings.

Pythons urllib2 will do "the right thing" when common environment
variables such as `http_proxy` or `https_proxy` are set.

Most noteably this also removes the `reporthook` nested function and
removes a lot of code from the `download_file` function.