extend sanity_check_paths for libraries
A new key (`'libs'`) is introduced potentially holding a list of maps
specifying libraries to be checked in a generic way.
Each library map must have a `'name'` and a `'kind'` while the latter
must either be `'shared'` or `'static'`. The name is only the library
name without path and suffix.

Example:
To test for a library `libmy.so`, which might either be in `lib/`
or in `lib64/`:

    sanity_check_paths = {
      'libs': [{'name': 'libmy', 'kind': 'shared'}]
    }

This will look for the following files:

    lib/libmy.so
    lib/libmy.dyn
    lib64/libmy.so
    lib64/libmy.dyn

This is done by generating a tuple of these, which is appended to the
`sanity_check_paths['files']` list.

Signed-off-by: Torbjörn Klatt <opensource@torbjoern-klatt.de>