Make Intel compiler + IntelMPI use mpiifort etc.
The combination of Intel compilers and IntelMPI should always use
mpiifort, mpiicc, mpiicpc and not mpif90, mpicc, mpicxx.

The main reason is that for Fortran90 code that uses
 USE mpi
mpif90 finds the gfortran mpi.mod and not the ifort one, this regardless
of having I_MPI_F77/F90/FC set to ifort or not.

Without this patch building of any fortran mpi code that uses mpi.mod or
any of the other Fortran modules fails.

It is also a recommendation by Intel to use mpiicc, mpiicpc, and
mpiifort for this combination and not to use mpif90 with I_MPI_F90 or
similar.