Add ec parameter to expand_toolchain_load().
So it can expand depending on the type of easyconfig.
In the module naming scheme used by Compute Canada
we need this because we like compiler-only toolchains
such as GCC and iccifort to be exposed and extend MODULEPATH,
and not be expanded, but higher-level toolchains to be hidden and
expanded.