Adjust testcase to the use of "module show".
The "if" logic works correctly now but it really tested for the modules
to exist in $HOME during the testcase which is not appropriate.

The reason the test worked previously is that the MODULEPATH extension
was taken unconditionally -- however that does not work if the
syntax is more complex as in e.g.
prepend_path("MODULEPATH", pathJoin(os.getenv("HOME"), "Public/easybuild/modules/Compiler/GCCcore/6.3.0"))
without parsing this was literally taken as the string
'pathJoin(os.getenv("HOME"), "Public/easybuild/modules/Compiler/GCCcore/6.3.0")'