Remove extra space in easyconfig.py update function.
The extra space added to self[key] in the update function should be removed.

It cases repeated calls to update the same string to have double " " in
them.

This causes problems in the namd easyblock later on when the string is
split and joined with "-", resulting in "xx-yy--zz" instead of
"xx-yy-zz".