Add iterate_builddependencies easyconfig option.
This is a list of list of buildependencies. It gets iterated on top
of the common builddependencies. This will allow e.g. to build with
multiple Python versions. A toy test is included.

The "prepare" step is now fully iterated which needed a few adjustments.
Hopefully this doesn't break anything.