Add support for GitHub token file to store a GitHub token.
Add support for GitHub token file to store a GitHub token as an
alternative to storing the GitHub token in a keyring.

The GitHub token file should be stored at:

  ~/.config/easybuild/github_token

and should only be readable by the owner of the file, which
should be the same user than the invoker of easybuild (same
permissions as private SSH keys).