Adding sorting to rst output of list_toolchains to match txt and none
When requesting --output-format=rst, the list of available toolchains was
not being sorted.  That output will be used to create a table of available
toolchains for the documentation and needs to be sorted.  I wrapped the
list of keys in `sorted()` to match the ouput of `eb --list-toolchains`
with no options and when `txt` is requested.