Deprecate failure to resolve a template value
A failure in template resolving is currently hidden in the log although
it is likely due to a bug in the code (either EasyBuild, EasyBlock or EasyConfig)
This may lead to cryptic error messages later on or misbehave silently
which is even worse.
To not break builds completely deprecate continuing for now and promote
to a hard error later.