Allow amending EC params which are not the default
Overwriting parameters which are not in the EC and not one of a few standard ones is currently impossible.
However it is useful when e.g. EasyBlocks introduce new params where the user wants to overwrite the default
value without changing the EC manually.
This change allows this by simply appending the new parameter when it is not in the default list.