Add module-write hook
This hook will be called anytime a module is written, so even for the
fake modules used for sanity checks etc.
It gets passed the easyblock as the first argument and the filename and
content to be written as the next 2 and may return a new text to be used
for the module file. If nothing or None is returned the original text
will be used.