Add option to ignore failing test step
Users can now use --ignore-test-failure instead of --skip-test-step
which will run the tests but does not abort on error and just prints it instead.

Using both options doesn't make sense and is hence disallowed and a
warning is show to make users move to ignoring failures instead of
skipping the test completely.