Avoid checking executable rights of subdirs of temp dirs already checked
`run_cmd` is one of the slowest things we can do as it forks the whole process.
In the check when setting a temp dir we run a dummy file to see if the
folder is actually executable. This means running this check at least
twice per test case and another time for each init_config call in the
test amounting to e.g. 263 calls in a single test (test_compiler_dependent_optarch)

So cache the results and don't repeat the run_cmd if it already
succeeded for a parent folder.
Likely no change at all in regular mode but really helps for the tests