Add an --extra-source-urls CLI option to fetch sources from additional URLs.
Add the --extra-source-urls CLI option, a comma separated list of
URLs that EasyBuild will fetch sources from. It replaces the hard-coded
EASYBUILD_SOURCES_URL, but keeps it as a default value. Also add --extra-source-urls-priority
CLI option is used to determine if these URLs are to be queried first or last. Possible
options are "primary" and "backup".

Co-authored-by: Alexander Grund <Flamefire@users.noreply.github.com>