Ignore request for external module (meta)data when no modules tool is active
E.g. for `--fetch` the modules tool is explicitly set to
`None`/`NoModulesTool` to allow running on e.g. laptops to fetch sources.
This however fails for external module dependencies in which it tries to
resolve them using the modules tool which fails with `NotImplementedError`.

Simply ignore the request in that case as-if the module metadata wasn't found.

Fixes #4298