Deduplicate warnings&errors found in logs and add initial newline in output
It isn't useful to report duplicate errors multiple times. And e.g.
configure may report a wrong option at the start and end.
Hence deduplicate it.
Also change the logged warning/error to include a newline & tab in front
of each element which makes it nicer to read. E.g.:
> build failed (first 300 chars): Found 1 error(s) in command output:
>	configure: WARNING: unrecognized options: --with-foo, --with-bar (took 2 mins 39 secs)