use source toolchain version when passing only `--try-toolchain`
`process_software_build_specs` ialways adds a `toolchain` key
even when only the name of the toolchain is available.
This makes the code in `tweak` ignore the code path in using
`toolchain_name` and/or `toolchain_version` with the fallback to using
the values of the source toolchain.
This in turn leads to failures further down when a value of `None` is
used where an actual version is expected.

Fix by only adding the `toolchain` key when we have both the name and
version and hence a complete, valid toolchain spec.