Cleanup & Speedup environment checks
`os.environ` is a dict hence the check for existance and access can
usually be combined into a single access of the key avoiding the
repeated lookup.
Especially for `os.getenv` which already returns `None` when the key
doesn't exist.
Also use `.pop(key, None)` to remove a key without checking for it's
existance first to avoid C&P errors