Eliminate use of distutils.version.StrictVersion
distutils was removed in Python 3.12. The only reason EasyBuild uses
StrictVersion is that it orders beta/rc versions before the released
version, unlike LooseVersion. E.g. 5.0.0-beta < 5.0.0 (but > for
LooseVersion).

So a new method
`is_earlier_or_prerelease(self, other, markers)`
was added to LooseVersion to handle that particular case.

Addresses part of #3963