`get_software_libdir`: return dir if it's only 1 with libs
There are situations e.g. with recent Gentoo where all 64-bit libraries
are under `lib64`, but `lib` is seperate and used for other files and
directories (e.g. `lib/python`), and for 32-bit libraries.

`get_software_libdir` would fail for such. To avoid this:

if `only_one` is `True` (default), *and* no specific file (`fs`) is
specified (also default), *and* `lib` and `lib64` both exist and are
separate, and only one of them has libraries (shared or static),
it returns that one directory.