Fix resolved values in case of failure
When the value cannot be resolved usually the original value is
returned. However if it contains `%` signs there will be an escape step
and that escaped value is returned.
That makes it impossible to have values which can only later be
resolved, like `cd %(startdir)s` in extensions which would become
`cd %%(startdir)s` on the first resolve-attempt.