fix `copy_file` with a folder as the target
This use case comes up with e.g. `--copy-ec /tmp`.
The existing code will see that `/tmp` exists and has a different UID
and hence will use `shutil.copy_file` which expects the target to not be
a folder. It hence fails with "[Errno 21] Is a directory"
Use `shutil.copy` as was likely intended anyway.