verify checksums with `--fetch`
After `--fetch` finishes it is reasonable to expect that the build won't
fail due to missing or wrong sources.
However putting the checksum-step into the source-step skips over the
verification and one would need to use `--stop=source` which does a lot
more than necessary, e.g. creating build dirs and environment
variables.

Move the checksums step into the fetch step and add a test for that.