Fix findPythonDeps for cyclic dependencies
There might be situations where the same dependency comes up multiple times
which leads to a cycle in the detection.
That might be due to a cyclic dependency or just some "unfortunate" structure
between packages.
Remove the recursive call leading to a possible error and maintain a set of packages
checked already to avoid checking any multiple times.
This makes sure the function terminates eventually.