Let jobs tweak easyconfigs themselves
This can be accomplished by tweak() optionally also returning a dict which
maps the tweaked easyconfig to the original version. Then the job can
run `eb ... <original_easyconfig.eb> --try-*` and that original
easyconfig will be retweaked in the job itself.

If the easyconfig passed to the job is not tweaked, then `--try-*` is
*not* passed down (so, with `--robot`, some jobs will have `--try-*`
and some don't).

This removes the requirement of a shared tmpdir with `--job --try-*`.

Fixes #1355