Allow templates in `custom_paths` & `custom_commands` sanity-check arguments
E.g. for Python packages we (almost) always want to check for
`lib/python%(pyshortver)s` and potentially add additional paths from
EasyConfigs.
Currently a workaround is used that sets the parameter in the easyconfig
when it isn't set already which is against the semantics of `enhance_sanity_check`.
This change allows this in a trivial way.