Add explicit property for defaulted versions
Currently there is no way to detect whether a version was defaulted.
The current default value ('0.0.0') is a valid value for the versions
list and the appropriate default comparator for that should be `==`.
However such a version is (mis-)detected as being the default and the
resulting default comparator is `>=` which is less restrictive than intended.

Fix by introducing a property set only for the default version.