EasyBlock: Allow to pass logfile from other EasyBlock
In the case an EasyBlock creates another instance of an EasyBlock (for
example when dealing with Bundles), each instance will create new
log file, which may stay open and cause issues down the line. To
prevent this, add an option to the EasyBlock constructor to pass a
logfile from another EasyBlock.

Signed-off-by: Jan André Reuter <j.reuter@fz-juelich.de>