Fix tests
Tests expected output of the rpath Python script to be space separated
but now it is line separated.
Also the output does not include the name of the array anymore.

To aid with debugging a helper method is introduced and the output is
compared as a list with one line each.