show readable error message when applying patch without (extracted) source
When there is no source `self.src` is an empty list which leads to a
rather generic error message. Improve that by showing that there was no
source to apply the patch to.

For extensions `self.src` is set to a string as only a single source is
supported.
Accessing `self.src[0].['finalpath']` leads to an error
> TypeError: string indices must be integers
This happens when the source didn't got extracted so `self.ext_dir` and
hence `beginpath` will be `None`.
Make the error show that the source was not extracted.