Disable exit code check when running "module show"
The "module show" command is used by ModulesTool "exist" method to test
the existence of a modulefile. Running "module show" against a
non-existing modulefile will make modulecmd.tcl script to produce a
non-zero exit code starting Modules v5.5.

With this change, the exit code check when running "module show" is
disabled. It should not harm the module existence check as it relies
on the analysis of the obtained stderr output.

Note that Lmod (as of version 8.7.55) does not set a non-zero exit code
when running "module show" against a non-existing modulefile when
LMOD_QUIET environment variable is set. If this behavior is changed in
the future, this commit will also benefit to Lmod.