Ignore unmodified files when using `--new-pr` or `--update_pr`
When including an unmodified EasyConfig with `--new-pr` an error is
shown that a commit message is required because this EC is modified
which is not the case.

Adjust the `copy_*` functions to ignore any unmodified file.
This especially ommits them in the `file_info` struct returned that is
then used to determine commit message/title etc.

Fixes #4751