Fix errors due to parallel property being unset
Move `set_parallel` to `post_init` which matches better than the `ready_step`.
That is described:
> creating build dir, resetting environment
> Verify if all is ok to start build.

None of that matches the behavior of a `set_parallel` call as opposed to
> Run post-initialization tasks.

When handling potentially additional limitations in extensions we need
to check if we have `parallel` set already before applying it.
This should only be the case for tests which take shortcuts in the logic
such as skipping `post_init`.