Avoid processing the same EasyConfig multiple times
When passing the same EasyConfig file multiple times on the commandline
it will be parsed once and put into a cache.
Every future call will get the same `EasyConfig` instance so any
modification to it will be reflected in all future uses of supposedly
"freshly parsed EasyConfigs".
This leads to confusing failures when trying to build the same file twice
which is likely a mistake anyway, especially when one file is a symlink
to another one passed too.
So just make sure each physical file is parsed only once.