Allow test case filtering by class name
It sometimes is useful to run either all tests from a class or a
specific test from a class where the name might also be used in another
class.
Use the common convention of '<class>.<test>' and match aginst that
allowing e.g. "TypeCheckingTest.test_check_type" as a filter.