Fix quoting of command in setup shell script
Avoid that variables are resolved when the `echo` is run.
`shlex.quote` outputs a properly escaped string so nothing will be
interpreted by the shell. It uses single quotes for that.
Putting that inside double quotes means single quotes are used literally
and and any variable will be resolved during the echo as shlex.quote
put them in single quotes to avoid that.

So end the double-quotes and append the shlex-quoted string.
To make it look neither add single quotes around it.