Make specifying `exts_defaultclass` optional
It is NOT required when all extensions have either custom easyblocks or
an `easyblock` key in their option dict.
In those cases `exts_defaultclass` is redundant making it harder to
write the EasyConfig.
Check for a missing easyblock when actually using it so the error will
be just a bit later but still (almost) the same.