Avoid unecessary warning about skipping checksum verification of Git repos
Some EasyConfigs use `None` as the checksum when using Git checkouts
hence skipping the checksum verification.
The deprecation warning for those is not necessary and potentially
confusing as it is not skipping because of Python < 3.9 but because
there wasn't a checksum on purpose.
Only show the warning if a checksum would actually be used.