Fix reporting of step name
The step method is a method returning the actual method.
The `Bundle` easyblock uses a `getattr` call there instead of the method name.

As the reflection looks at local name(s) it will find the `getattr`
instead of the actual method.

Access the `__name__` property of the returned function instead so we
will always get the right name no matter how the lambda/method is defined.