Don't log variables (un)set when  cleaning up fake module
Unloading a module (and its dependencies) changes a massive amount of
variables.
As the changes done by *unloading* the fake module aren't really relevant
for the remaining build (it is usually done at the end) do not log them.

This avoids filling up the part of the log uploaded with test reports
with irrelevant information.