Check module extensions support within Tcl modulefiles
Instead of making EasyBuild check if module extensions are supported by
module tool to generate the proper Tcl modulefiles, move this check
directly within modulefile code, like done with Lua modulefiles.

It means module extensions code is always written in modulefile, but a
condition guards it, to execute this code only if the version of the
module tool evaluating the modulefile supports it.

So instead of checking the version of the module tool used by EasyBuild,
the check is made on the module tool that evaluates the generated
modulefiles.

Even with this new approach, the "supports_extensions" ModuleTool state
is kept to cope with older ModuleTool, like EnvironmentModulesC.

Signed-off-by: Xavier Delaruelle <xavier.delaruelle@cea.fr>