Handle post_install_patches and post_install_msgs for extensions
Those EasyConfig parameters were silently ignored.
Run the appropriate steps in the default implementation of
`post_install_extension` similar to postinstall_cmds and add a test.
Using the new name for `post_install_patches` in the test reveals the need
to handle that explicitely as a raw dict is used.