Handle extracting the same file to the same destination
When a source is added multiple times the content after the 2nd
extraction will be the same and no new files/folders are detected.
So `extract_file` would return the parent folder.
However if there was a clear result after the first extraction the same
result should be returned.

This works trivially until something else is extracted to the same
folder between the 2 extraction operations.